requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14
psycopg2-binary==2.9.9
gunicorn>=21.2
python-telegram-bot>=20.0,<21.0
python-dotenv>=1.0
drf-yasg>=1.21
requests
python-dateutil>=2.8.2
celery>=5.3
redis>=5.0
django-celery-beat>=2.5.0
whitenoise
Pillow
# django-jazzmin
pyjwt
python-dotenv
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn[standard]